% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignOut.R
\name{asGoogleSignOut}
\alias{asGoogleSignOut}
\title{Use an HTML element for Google sign-out}
\usage{
asGoogleSignOut(tag)
}
\arguments{
\item{tag}{An \link[htmltools:builder]{htmltools::tag} object.}
}
\description{
Clicks on marked elements will trigger a Google sign-out.
}
\details{
A sign-out will clear all \code{googleSignInButton()} inputs, and prevent
automatic sign-in from happening afterwards.
}
\examples{
library(shiny)

# Use a regular action button to trigger a sign out
asGoogleSignOut(actionButton("sign_out", "Sign out"))
}
\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/automatic-sign-in-sign-out#sign-out}
}
}
\seealso{
Other sign-out methods: 
\code{\link{sendGoogleSignOut}()}
}
\concept{sign-out methods}
