% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{google_public_key}
\alias{google_public_key}
\title{Get Google's JWK public key for verifying signatures}
\usage{
google_public_key()
}
\value{
An OpenSSL RSA public key object. See \code{\link[openssl:keygen]{openssl::rsa_keygen()}}.
}
\description{
Get Google's JWK public key via an API call. The result is cached according
to the \code{Cache-Control} header with the \link{httc} package.
}
\keyword{internal}
