% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{googleSignInApp}
\alias{googleSignInApp}
\title{A demo Shiny app for using Sign In With Google}
\usage{
googleSignInApp(client_id)
}
\arguments{
\item{client_id}{A string containing the client ID of your Google web app.
See Details for acquiring one.}
}
\description{
A demo Shiny app for using Sign In With Google
}
\section{Client ID}{
In order to use Sign In With Google in Shiny, you need to register a client
in a Google Cloud Platform project and set it up for OAuth use. Google
provides details for the setup steps
(\url{https://developers.google.com/identity/gsi/web/guides/get-google-api-clientid}).
}

\seealso{
Individual components: \code{\link[=useGoogleSignIn]{useGoogleSignIn()}}, \code{\link[=googleSignInUI]{googleSignInUI()}} and
\code{\link[=googleSignInServer]{googleSignInServer()}}.
}
