% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{googleSignInButton}
\alias{googleSignInButton}
\title{Create HTML for a Sign In With Google button}
\usage{
googleSignInButton(inputId, options = list())
}
\arguments{
\item{inputId}{The \code{input} slot that authentication information will be
available in. Note that while you can have several buttons in an app, the
\code{inputId} of the first one found will determine the name of the Shiny
\code{input} where authentication information will be passed.}

\item{options}{A list of customization options to control the appearance of
the button. \verb{data-} will be automatically appended to the names. See \href{https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin}{GSI reference}
for possible values.}
}
\description{
Creates the HTML element for displaying a Sign In With Google button. Most of
the time you probably don't want to use this directly, but through the UI
module \code{googleSignInUI()} instead. That allows the server to decode and
verify the token received from the authentication process.
}
\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/display-button#html}
\item \url{https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin}
}
}
