% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{googleSignInServer}
\alias{googleSignInServer}
\title{Server side for a Sign In With Google module}
\usage{
googleSignInServer(id, client_ids)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{client_ids}{A character vector of recognized client IDs for your app.}
}
\description{
The module provides a reactive return value containing a decoded and verified
Google ID token, or NULL. See \code{\link[=gsi_verify_credential]{gsi_verify_credential()}} for details.
}
\seealso{
\code{\link[=googleSignInUI]{googleSignInUI()}} for the module UI part.
}
