% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{googleSignInUI}
\alias{googleSignInUI}
\title{User interface for a Sign In With Google module}
\usage{
googleSignInUI(id, options = list())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{options}{A list of customization options to control the appearance of
the button. \verb{data-} will be automatically appended to the names. See \href{https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin}{GSI reference}
for possible values.}
}
\description{
User interface for a Sign In With Google module
}
\details{
While you \emph{can} include several of these in an app, only the first
one will ever receive input from the sign in button due to the way Sign In
With Google is set up on the JavaScript side.
}
\seealso{
\code{\link[=googleSignInServer]{googleSignInServer()}} for the module server part.
}
