% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{gsi_user_info}
\alias{gsi_user_info}
\title{Get user details from a Google ID token}
\usage{
gsi_user_info(credential)
}
\arguments{
\item{credential}{A list containing a decoded Google ID token.}
}
\value{
A list with fields describing basic user data for the Google user, or
\code{NULL} if given a \code{NULL} value as \code{credential}.
}
\description{
Get user details from a Google ID token
}
\seealso{
\code{gsi_verify_credential()} for decoding and verifying an encoded
Google ID JWT.
}
