% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{gsi_verifier}
\alias{gsi_verifier}
\title{Create a verifier for verifying ID tokens}
\usage{
gsi_verifier(client_ids = character(), public_keys = NULL)
}
\arguments{
\item{client_ids}{A character vector of client IDs your app recognizes.}

\item{public_keys}{A list of public keys tried for signature verification,
or a function that returns them. Keys can be objects, or paths to read
with \code{\link[openssl:read_key]{openssl::read_pubkey()}}. If left \code{NULL}, fetch Google's current
public keys from their API when verifying.}
}
\description{
Create a verifier for verifying ID tokens
}
\examples{
gsi_verifier()
}
\seealso{
\code{gsi_verify()} for performing the verification.
}
