% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{gsi_verify}
\alias{gsi_verify}
\title{Decode and verify an encoded Google ID token}
\usage{
gsi_verify(verifier, token)
}
\arguments{
\item{verifier}{A verifier object. See \code{gsi_verifier()}.}

\item{token}{A string with an encoded Google ID \href{https://jwt.io}{JWT}.}
}
\value{
The decoded JWT payload. Signals an error if verification failed.
}
\description{
Decode and verify an encoded Google ID token
}
\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/verify-google-id-token}
}
}
\seealso{
\code{\link[=gsi_user_info]{gsi_user_info()}} for extracting user details from the decoded JWT.
}
