% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{gsi_verify_credential}
\alias{gsi_verify_credential}
\title{Decode and verify an encoded Google ID token}
\usage{
gsi_verify_credential(credential, client_ids)
}
\arguments{
\item{credential}{A string containing an encoded Google ID JWT.}

\item{client_ids}{A character vector of recognized client IDs for your app.}
}
\value{
\code{NULL} if verification failed, otherwise the decoded JWT payload.
}
\description{
Decode and verify an encoded Google ID token
}
\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/verify-google-id-token}
}
}
\seealso{
\code{\link[=gsi_user_info]{gsi_user_info()}} for extracting user details from the decoded JWT.
}
