% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignOut.R
\name{sendGoogleSignOut}
\alias{sendGoogleSignOut}
\title{Trigger a Google sign-out from server-side}
\usage{
sendGoogleSignOut(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The \code{shiny::session} object.}
}
\description{
Trigger a Google sign-out from server-side
}
\details{
A sign-out will clear all \code{googleSignInButton()} inputs, and prevent
automatic sign-in from happening afterwards.
}
\seealso{
Other sign-out methods: 
\code{\link{asGoogleSignOut}()}
}
\concept{sign-out methods}
