% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygsi-package.R
\docType{package}
\name{shinygsi-package}
\alias{shinygsi}
\alias{shinygsi-package}
\title{shinygsi: Integrate Sign In With Google with Shiny Apps}
\description{
Use Sign In With Google in your Shiny app to authenticate users, and get their basic information.
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
