% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{useGoogleSignIn}
\alias{useGoogleSignIn}
\title{Set up a Shiny app to use Sign In With Google}
\usage{
useGoogleSignIn(client_id, auto_prompt = TRUE, auto_select = TRUE)
}
\arguments{
\item{client_id}{A string containing the Google API client ID of your app.
See Details for acquiring one.}

\item{auto_prompt}{Logical. Should \href{https://developers.google.com/identity/gsi/web/guides/features}{the Google One Tap prompt} be displayed?}

\item{auto_select}{Logical. Should \href{https://developers.google.com/identity/gsi/web/guides/automatic-sign-in-sign-out}{automatic sign-in} be enabled?}
}
\description{
Call this function at least once in your Shiny app UI to set up the necessary
elements for Sign In With Google to work.
}
\section{Client ID}{
In order to use Sign In With Google in Shiny, you need to register a client
in a Google Cloud Platform project and set it up for OAuth use. You can find
details for the process in \href{https://developers.google.com/identity/gsi/web/guides/get-google-api-clientid}{Google's setup guide}.
}

\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/display-button#html}
\item \url{https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_id_g_id_onload}
}
}
\seealso{
Other module functions: 
\code{\link{googleSignInServer}()},
\code{\link{googleSignInUI}()}
}
\concept{module functions}
