% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{googleSignInServer}
\alias{googleSignInServer}
\title{Server side for a Sign In With Google module}
\usage{
googleSignInServer(id, client_ids)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{client_ids}{A character vector of recognized client IDs for your app.}
}
\description{
Handles decoding and verification of the authentication response received
from the linked UI part, \code{googleSignInUI()}. Provides a reactive return value
containing a list with the verified Google ID token's payload, or \code{NULL}.
}
\details{
The server module decodes and verifies the encoded Google ID JWT received
from the UI module upon succesful completion of the authentication flow. See
\code{\link[=gsi_verify_credential]{gsi_verify_credential()}} for details about the verification process.
}
\seealso{
\code{\link[=gsi_user_info]{gsi_user_info()}} for extracting user details from the return value.

Other module functions: 
\code{\link{googleSignInUI}()},
\code{\link{useGoogleSignIn}()}
}
\concept{module functions}
