% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygsi-package.R
\docType{package}
\name{shinygsi-package}
\alias{shinygsi}
\alias{shinygsi-package}
\title{shinygsi: Integrate Sign In With Google with Shiny Apps}
\description{
Use Sign In With Google in your Shiny app to authenticate users, and get their basic information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikmart.github.io/shinygsi}
  \item \url{https://github.com/mikmart/shinygsi}
  \item Report bugs at \url{https://github.com/mikmart/shinygsi/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
