% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygsi-package.R
\docType{package}
\name{shinygsi-package}
\alias{shinygsi}
\alias{shinygsi-package}
\title{shinygsi: Use Sign In With Google in Shiny Apps}
\description{
A Shiny module to easily integrate authentication provided by
    Google Identity Services in your app. Features include a personalized
    sign-in button, the Google One Tap prompt, and automatic sign-in.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikmart.github.io/shinygsi}
  \item \url{https://github.com/mikmart/shinygsi}
  \item Report bugs at \url{https://github.com/mikmart/shinygsi/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
