% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSignIn.R
\name{useGoogleSignIn}
\alias{useGoogleSignIn}
\title{Set up a Shiny app to use Sign In With Google}
\usage{
useGoogleSignIn(client_id, auto_prompt = TRUE)
}
\arguments{
\item{client_id}{A string containing the client ID of your Google web app.
See Details for acquiring one.}

\item{auto_prompt}{Logical. Should the Google One Tap prompt be displayed?}
}
\description{
Call this function at least once in your Shiny app UI to set up the necessary
elements for Sign In With Google to work.
}
\section{Client ID}{
In order to use Sign In With Google in Shiny, you need to register a client
in a Google Cloud Platform project and set it up for OAuth use. Google
provides details for the setup steps
(\url{https://developers.google.com/identity/gsi/web/guides/get-google-api-clientid}).
}

\references{
\itemize{
\item \url{https://developers.google.com/identity/gsi/web/guides/display-button#html}
\item \url{https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_id_g_id_onload}
}
}
\seealso{
The module parts: \code{\link[=googleSignInUI]{googleSignInUI()}} and \code{\link[=googleSignInServer]{googleSignInServer()}}.

Other module functions: 
\code{\link{googleSignInServer}()},
\code{\link{googleSignInUI}()}
}
\concept{module functions}
